Add in maidenhead grid locator conversions
It's currently only Grid<6>, but maybe I'll add Grid<4> soon. This is
handy for taking ham radio geospatial info in the form of grids and
using them with, say, haversine or azimuth computations. This isn't
super useful for high precision stuff like local tangent plane
operations since the grids are so big, but it may be handy to someone
out there - and I had it already written.